orchestrator/requirements.txt
# Orchestrator Agent Dependencies

# Core async and web framework dependencies
asyncio-mqtt==0.16.2
aiohttp==3.9.5
aiofiles==23.2.1

# FastAPI integration (to communicate with UX analyzer)
httpx==0.27.0
requests==2.32.3

# Configuration and environment management
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4

# Logging and monitoring
structlog==24.1.0
colorlog==6.8.2

# Task scheduling and orchestration
celery==5.3.6
redis==5.0.7
schedule==1.2.2

# Azure DevOps integration (for coder agent)
azure-devops==7.1.0b4
azure-identity==1.17.1

# GitHub integration
PyGithub==2.3.0
gitpython==3.1.43

# AI/ML integration
google-generativeai==0.7.2
openai==1.35.15

# Database (for storing orchestration history)
sqlalchemy==2.0.31
# sqlite3 is built into Python - no need to install

# Notification integrations
slack-sdk==3.29.0
sendgrid==6.11.0

# Utilities
click==8.1.7
rich==13.7.1
tabulate==0.9.0
pathlib2==2.3.7

# Testing (development)
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0

# Security
cryptography==42.0.8
python-jose==3.3.0

# Development tools
black==24.4.2
flake8==7.1.0
mypy==1.10.1
